settings: reorganise
