+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./App.css";
 import { DataProvider } from "./data-lib/context/DataProvider";
 import useInBound from "./hook/useInBound";
-import YearPlanner from "./comp/YearPlanner/YearPlanner";
+import YearPlanner from "./comp/1-YearPlanner/YearPlanner";
 
 function App() {
 const { prepopData } = useInBound({});
